Upgrade menu with parsing for complex commands, find prototype
Working towards a set of commands that take options and 
multiple arguments.  The original two find commands 
are now accessed with 'find1' and 'find2', and now a new 
find command can be accessed with 'find -h' for help, 
though there are no further functions.
Also, the new find command has errors if there isn't a 
second command or argument, which needs to be fixed 
by checking the number of command array elements 
before testing their contents (which may not exist).

cli_finds
Prototyping the find cli command.
Added a setup function to store the rows and columns of 
the screen in screenRows and screenCols memory location.

browse function will be for testing scrolling up and down 
through memory records.  The test is about whether 
printing is fast enough to print every line on the screen 
in a instant to simulate scrolling up.