Browse function draft
In support of the find command, I need a way to fluently 
browse through lists of things, such as memory or objects or people.
The browse command is moving towards a way to allow scrolling 
through lists, selecting an item, and then taking actions on it.

To make browse generic, it should probably accept an array and output a number.
Then it can be paired with another function which either takes appropriate 
action, or gives additional options for actions to take.

It would be nice if it allowed scrolling through 5 or 10 at a time when 
pressing the right or left arrow keys.

displayBuffer offloads the repetative plotting for browse.
The line numbers should be adjusted to be aligned.