Cleanup tabs, add Copy and Slurp features
browse will probably need to be made more generic, 
then wrapped for browsing memory, computers, 
people, objects, etc.

To do that, the actions part of browse starting on line 
125 will probably need to be extracted and 
moved to a different function.  Browse could 
just return the cursor index, and that could 
be used to feed the next menu.