Split browse() into browseMemory() and browseArray()
Moving browse() from a proof of concept to a
more reusable function.
number browseArray(array inputArray) provides scrollable
menu of the inputArray, and when an entry is selected,
the index is returned.  This is intended to be simple
to reuse in the code on any array.

browseMemory() is a wrapper function just to maintain
the previous functionality with the new generic
browseArray() function.